Create and get key all working
