+++ typescript/infra/scripts/utils.ts
@@ -8,7 +8,7 @@ import {
 } from '@abacus-network/sdk';
 import { types } from '@abacus-network/utils';
 import { RouterConfig } from '@abacus-network/deploy';
-import { KEY_ROLE_ENUM } from '../src/agents';
+import { KEY_ROLE_ENUM } from '../src/agents/roles';
 import {
   ALL_ENVIRONMENTS,
   AgentConfig,
