+++ typescript/infra/src/agents/aws/relayer-user.ts
@@ -1,5 +1,5 @@
 import { ChainName } from '@abacus-network/sdk';
-import { KEY_ROLE_ENUM } from '..';
+import { KEY_ROLE_ENUM } from '../roles';
 import { AgentConfig } from '../../config';
 import { AgentAwsKey } from './key';
 import { AgentAwsUser } from './user';
