+++ typescript/infra/src/agents/aws/validator-user.ts
@@ -5,7 +5,7 @@ import {
   ListBucketsCommand,
   PutBucketPolicyCommand,
 } from '@aws-sdk/client-s3';
-import { KEY_ROLE_ENUM } from '../../agents';
+import { KEY_ROLE_ENUM } from '../roles';
 import { AgentConfig } from '../../config';
 import { AgentAwsKey } from './key';
 import { AgentAwsUser } from './user';
