+++ typescript/infra/src/agents/index.ts
@@ -7,67 +7,9 @@ import { ensure0x, execCmd, strip0x } from '../utils/utils';
 import { AgentGCPKey, fetchAgentGCPKeys } from './gcp';
 import { AgentAwsKey } from './aws/key';
 import { ChainAgentConfig, CheckpointSyncerType } from '../config/agent';
-import { AgentKey, identifier } from './agent';
+import { identifier } from './agent';
 import { AgentAwsUser, ValidatorAgentAwsUser } from './aws';
-
-export enum KEY_ROLE_ENUM {
-  Validator = 'validator',
-  Checkpointer = 'checkpointer',
-  Relayer = 'relayer',
-  Deployer = 'deployer',
-  Bank = 'bank',
-  Kathy = 'kathy',
-}
-
-export const KEY_ROLES = [
-  KEY_ROLE_ENUM.Validator,
-  KEY_ROLE_ENUM.Checkpointer,
-  KEY_ROLE_ENUM.Relayer,
-  KEY_ROLE_ENUM.Deployer,
-  KEY_ROLE_ENUM.Bank,
-  KEY_ROLE_ENUM.Kathy,
-];
-
-export function getKey<Networks extends ChainName>(
-  agentConfig: AgentConfig<Networks>,
-  role: KEY_ROLE_ENUM,
-  chainName?: Networks,
-  suffix?: Networks | number,
-): AgentKey<Networks> {
-  if (agentConfig.aws) {
-    return new AgentAwsKey(
-      agentConfig,
-      role,
-      chainName,
-      suffix,
-    )
-  } else {
-    return new AgentGCPKey(
-      agentConfig,
-      role,
-      chainName,
-      suffix,
-    );
-  }
-}
-
-export function getAllKeys<Networks extends ChainName>(agentConfig: AgentConfig<Networks>): Array<AgentKey<Networks>> {
-  return KEY_ROLES.flatMap((role) => {
-    if (role === KEY_ROLE_ENUM.Validator) {
-      // For each chainName, create validatorCount keys
-      return agentConfig.domainNames.flatMap((chainName) =>
-        [...Array(agentConfig.validatorSets[chainName].validators.length).keys()].map((index) =>
-          getKey(agentConfig, role, chainName, index),
-        ),
-      );
-    } else if (role === KEY_ROLE_ENUM.Relayer) {
-      return agentConfig.domainNames
-          .map((chainName) => getKey(agentConfig, role, chainName));
-    } else {
-      return [getKey(agentConfig, role)];
-    }
-  });
-}
+import { KEY_ROLES, KEY_ROLE_ENUM } from './roles';
 
 async function helmValuesForChain<Networks extends ChainName>(
   chainName: Networks,
